requirements.txt
ccxt
ta
pandas
numpy
scikit-learn
joblib
tweepy
transformers
torch
python-dotenv